backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
python-multipart
python-docx
python-pptx
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic
pydantic-settings
google-generativeai
